Commit Before Destruction
