+++ .vscode/settings.json
@@ -1,0 +1,36 @@
+{
+          "workbench.colorTheme": "Gruvbox Dark (Hard)",
+          "git.enableSmartCommit": true,
+          "git.confirmSync": false,
+          "gitlens.currentLine.enabled": false,
+          "gitlens.hovers.currentLine.over": "line",
+          "gitlens.codeLens.enabled": false,
+          "editor.minimap.enabled": false,
+          "gitlens.blame.heatmap.location": "left",
+          "explorer.decorations.colors": false,
+          "explorer.decorations.badges": false,
+          "cmake.allowCommentsInPresetsFile": true,
+          "cmake.useCMakePresets": "always",
+          "cmake.configureOnOpen": true,
+          "cmake.options.statusBarVisibility": "visible",
+          "editor.fontFamily": "Victor Mono Regular",
+          "editor.fontWeight": "normal",
+          "workbench.iconTheme": "material-icon-theme",
+          "editor.fontSize": 20,
+          "workbench.editor.showTabs": "none",
+          "explorer.confirmDelete": false,
+          "cmake.showOptionsMovedNotification": false,
+          "editor.tabSize": 10,
+          "C_Cpp.intelliSenseEngine": "disabled",
+          "cmake.showConfigureWithDebuggerNotification": false,
+          "editor.renderLineHighlight": "none",
+          "breadcrumbs.enabled": false,
+          "editor.stickyScroll.enabled": false,
+          "workbench.activityBar.location": "top",
+          "editor.guides.indentation": false,
+          "notebook.output.fontSize": 10,
+          "editor.formatOnSave": true,
+          "workbench.layoutControl.enabled": false,
+          "editor.inlayHints.fontSize": 20,
+          "editor.inlayHints.enabled": "offUnlessPressed"
+}
