+++ compile_commands.json
@@ -619,21 +619,21 @@
 },
 {
   "directory": "C:/Users/sansk/OneDrive/Pictures/jkrgui/out/build/x64-debug",
-  "command": "C:\\PROGRA~1\\LLVM\\bin\\clang.exe -DTRACY_ENABLE -DVK_USE_PLATFORM_WIN32_KHR -D_CRT_SECURE_NO_WARNINGS -D_SDL_MAIN_HANDLED -IC:/Users/sansk/OneDrive/Pictures/jkrgui/application -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/ThreeD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/TwoD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/TextD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vulkan -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor/stbi -I/include} -IC:/Users/sansk/OneDrive/Pictures/jkrgui/libs -IC:/Users/sansk/OneDrive/Pictures/jkrgui/aJkrLua -IC:/Users/sansk/OneDrive/Pictures/jkrgui/vendor/lua -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor/Tracy/tracy -IC:/VulkanSDK/1.3.275.0/Include -O0 -std=gnu++20 -D_DEBUG -D_DLL -D_MT -Xclang --dependent-lib=msvcrtd -g -Xclang -gcodeview -flto=thin -o jkrgui\\CMakeFiles\\jkrgui.dir\\Renderers\\ThreeD\\glTF.cpp.obj -c C:\\Users\\sansk\\OneDrive\\Pictures\\jkrgui\\jkrgui\\Renderers\\ThreeD\\glTF.cpp",
-  "file": "C:\\Users\\sansk\\OneDrive\\Pictures\\jkrgui\\jkrgui\\Renderers\\ThreeD\\glTF.cpp",
-  "output": "jkrgui\\CMakeFiles\\jkrgui.dir\\Renderers\\ThreeD\\glTF.cpp.obj"
+  "command": "C:\\PROGRA~1\\LLVM\\bin\\clang.exe -DTRACY_ENABLE -DVK_USE_PLATFORM_WIN32_KHR -D_CRT_SECURE_NO_WARNINGS -D_SDL_MAIN_HANDLED -IC:/Users/sansk/OneDrive/Pictures/jkrgui/application -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/ThreeD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/TwoD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/TextD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vulkan -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor/stbi -I/include} -IC:/Users/sansk/OneDrive/Pictures/jkrgui/libs -IC:/Users/sansk/OneDrive/Pictures/jkrgui/aJkrLua -IC:/Users/sansk/OneDrive/Pictures/jkrgui/vendor/lua -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor/Tracy/tracy -IC:/VulkanSDK/1.3.275.0/Include -O0 -std=gnu++20 -D_DEBUG -D_DLL -D_MT -Xclang --dependent-lib=msvcrtd -g -Xclang -gcodeview -flto=thin -o jkrgui\\CMakeFiles\\jkrgui.dir\\Renderers\\ThreeD\\Shape3D.cpp.obj -c C:\\Users\\sansk\\OneDrive\\Pictures\\jkrgui\\jkrgui\\Renderers\\ThreeD\\Shape3D.cpp",
+  "file": "C:\\Users\\sansk\\OneDrive\\Pictures\\jkrgui\\jkrgui\\Renderers\\ThreeD\\Shape3D.cpp",
+  "output": "jkrgui\\CMakeFiles\\jkrgui.dir\\Renderers\\ThreeD\\Shape3D.cpp.obj"
+},
+{
+  "directory": "C:/Users/sansk/OneDrive/Pictures/jkrgui/out/build/x64-debug",
+  "command": "C:\\PROGRA~1\\LLVM\\bin\\clang.exe -DTRACY_ENABLE -DVK_USE_PLATFORM_WIN32_KHR -D_CRT_SECURE_NO_WARNINGS -D_SDL_MAIN_HANDLED -IC:/Users/sansk/OneDrive/Pictures/jkrgui/application -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/ThreeD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/TwoD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/TextD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vulkan -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor/stbi -I/include} -IC:/Users/sansk/OneDrive/Pictures/jkrgui/libs -IC:/Users/sansk/OneDrive/Pictures/jkrgui/aJkrLua -IC:/Users/sansk/OneDrive/Pictures/jkrgui/vendor/lua -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor/Tracy/tracy -IC:/VulkanSDK/1.3.275.0/Include -O0 -std=gnu++20 -D_DEBUG -D_DLL -D_MT -Xclang --dependent-lib=msvcrtd -g -Xclang -gcodeview -flto=thin -o jkrgui\\CMakeFiles\\jkrgui.dir\\Renderers\\ThreeD\\Shape3D_base.cpp.obj -c C:\\Users\\sansk\\OneDrive\\Pictures\\jkrgui\\jkrgui\\Renderers\\ThreeD\\Shape3D_base.cpp",
+  "file": "C:\\Users\\sansk\\OneDrive\\Pictures\\jkrgui\\jkrgui\\Renderers\\ThreeD\\Shape3D_base.cpp",
+  "output": "jkrgui\\CMakeFiles\\jkrgui.dir\\Renderers\\ThreeD\\Shape3D_base.cpp.obj"
 },
 {
   "directory": "C:/Users/sansk/OneDrive/Pictures/jkrgui/out/build/x64-debug",
   "command": "C:\\PROGRA~1\\LLVM\\bin\\clang.exe -DTRACY_ENABLE -DVK_USE_PLATFORM_WIN32_KHR -D_CRT_SECURE_NO_WARNINGS -D_SDL_MAIN_HANDLED -IC:/Users/sansk/OneDrive/Pictures/jkrgui/application -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/ThreeD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/TwoD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/TextD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vulkan -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor/stbi -I/include} -IC:/Users/sansk/OneDrive/Pictures/jkrgui/libs -IC:/Users/sansk/OneDrive/Pictures/jkrgui/aJkrLua -IC:/Users/sansk/OneDrive/Pictures/jkrgui/vendor/lua -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor/Tracy/tracy -IC:/VulkanSDK/1.3.275.0/Include -O0 -std=gnu++20 -D_DEBUG -D_DLL -D_MT -Xclang --dependent-lib=msvcrtd -g -Xclang -gcodeview -flto=thin -o jkrgui\\CMakeFiles\\jkrgui.dir\\Renderers\\ThreeD\\glTF_Model.cpp.obj -c C:\\Users\\sansk\\OneDrive\\Pictures\\jkrgui\\jkrgui\\Renderers\\ThreeD\\glTF_Model.cpp",
   "file": "C:\\Users\\sansk\\OneDrive\\Pictures\\jkrgui\\jkrgui\\Renderers\\ThreeD\\glTF_Model.cpp",
   "output": "jkrgui\\CMakeFiles\\jkrgui.dir\\Renderers\\ThreeD\\glTF_Model.cpp.obj"
-},
-{
-  "directory": "C:/Users/sansk/OneDrive/Pictures/jkrgui/out/build/x64-debug",
-  "command": "C:\\PROGRA~1\\LLVM\\bin\\clang.exe -DTRACY_ENABLE -DVK_USE_PLATFORM_WIN32_KHR -D_CRT_SECURE_NO_WARNINGS -D_SDL_MAIN_HANDLED -IC:/Users/sansk/OneDrive/Pictures/jkrgui/application -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/ThreeD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/TwoD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/jkrgui/Renderers/TextD -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vulkan -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor/stbi -I/include} -IC:/Users/sansk/OneDrive/Pictures/jkrgui/libs -IC:/Users/sansk/OneDrive/Pictures/jkrgui/aJkrLua -IC:/Users/sansk/OneDrive/Pictures/jkrgui/vendor/lua -IC:/Users/sansk/OneDrive/Pictures/jkrgui/ksaivulkan/Vendor/Tracy/tracy -IC:/VulkanSDK/1.3.275.0/Include -O0 -std=gnu++20 -D_DEBUG -D_DLL -D_MT -Xclang --dependent-lib=msvcrtd -g -Xclang -gcodeview -flto=thin -o jkrgui\\CMakeFiles\\jkrgui.dir\\Renderers\\ThreeD\\glTF_base.cpp.obj -c C:\\Users\\sansk\\OneDrive\\Pictures\\jkrgui\\jkrgui\\Renderers\\ThreeD\\glTF_base.cpp",
-  "file": "C:\\Users\\sansk\\OneDrive\\Pictures\\jkrgui\\jkrgui\\Renderers\\ThreeD\\glTF_base.cpp",
-  "output": "jkrgui\\CMakeFiles\\jkrgui.dir\\Renderers\\ThreeD\\glTF_base.cpp.obj"
 },
 {
   "directory": "C:/Users/sansk/OneDrive/Pictures/jkrgui/out/build/x64-debug",
