--- jkrgui/Renderers/ThreeD/glTF_base.cpp
@@ -1,2 +1,0 @@
-#include "glTF_base.hpp"
-
